requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.0
flask-session==0.6.0
flask-babel==4.0.0
pymysql==1.1.0
cryptography==42.0.5  # PyMySQL的加密支持
python-dotenv==1.0.1
flask-migrate==4.0.5  # 添加这行